+++ docs/conf.py
@@ -11,6 +11,8 @@ import sys
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 sys.path.insert(0, os.path.abspath('..'))
+sys.path.insert(0, os.path.abspath('../../examples'))
+
 
 # -- General configuration -----------------------------------------------------